Update cs1.md
In many programming languages, we need to declare the type of data a variable can store before using it. For example, in JavaScript, a variable named “age” which stores the value 16, has to be declared with the integer data type, by preceding the assignment statement with the word “int”, to not result in errors. It would look like this: int age = 16.

There are many other data types that can be used with variables. These types include numbers, Booleans, lists, and strings. 
The following is a list of the most common data types and what they represent:

|          Type           |                     Description                        |    example    |
|          :--:           |                       :----:                           |     :---:     | 
| String (or str or text) |  for characters, letters, numbers and symbols.         |    “43llo”    |
| Integer (or int)        |  for whole numbers.                                    |   443, 5, 66  |
| Float (or Real)         |  for numbers that contain decimal points, or fractions.|      55.6     |
| Boolean (or bool)       |  True/False                                            |        -      |

For further information, share the following article with students: https://www.bbc.co.uk/bitesize/guides/zwmbgk7/revision/3

In these levels, variables are used to name enemies, so that the hero knows which character to attack.