Update apcsp-big-data.md
Addition
Add to section: What is Big Data?
Over time, the way we store data has increased in scalability. That means that computers have become more advanced and storage costs have gone down, which allows people to store more and more information about the world. A spreadsheet that used to fit in a room-sized 4 megabyte hard drive used to be "big data". Now? Google is estimated to store 10+ exabytes of information. An exabyte is 1,000 petabytes... A petabyte is 1,000 terabytes... And a terabyte is 1,000 gigabytes! So around 1,000,000,000 gigabytes of data is stored and processed by Google. That's around 16 million modern video game installs!