Update cs2.md
Addition
Standard: AAP-3.B.3.
Add to section: Instructive Activity: Functions.
Then when the coach wants this sequence of actions to happen, she just calls out the name of the play: “Out-over-up!”. This practice of grouping different functions to build a program is commonplace in programming. The subdivision of a computer program into separate subprograms is called modularity.